requirements.txt
appdirs==1.4.4
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
distlib==0.3.1
Django==3.2.3
django-allauth==0.44.0
django-background-tasks==1.2.5
django-cleanup==5.2.0
django-compat==1.0.15
django-js-asset==1.2.2
django-jsonfield==1.4.1
django-mptt==0.12.0
django-mysql==3.11.1
filelock==3.0.12
idna==2.10
importlib-metadata==4.3.0
oauthlib==3.1.0
pbr==5.6.0
Pillow==8.2.0
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
stevedore==3.3.0
typing-extensions==3.10.0.0
urllib3==1.26.5
virtualenv==20.4.7
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
zipp==3.4.1
